constitution manager
