+++ package.json
@@ -12,7 +12,7 @@
 	"license": "ISC",
 	"devDependencies": {},
 	"dependencies": {
-		"@tableaubits/hang": "^1.2.0",
+		"@tableaubits/hang": "^1.3.5",
 		"@types/cors": "^2.8.10",
 		"@types/ejs": "^3.0.7",
 		"@types/express": "^4.17.11",
